refactor: MainElements 로 파일 분리
